publisher/requirements.txt
Flask==1.1.1
google-api-python-client==1.7.11
google-auth==1.11.2
google-cloud-pubsub==1.2.0